# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmergencyContact {
  id: number;
  name: string;
  phoneNumber: string;
  relationship: string;
  address: string;
  createdAt: string;
  updatedAt: string;
}

export interface EmergencyContactInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 3
     * @maxLength 40
     */
  phoneNumber: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  relationship: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  address: string;
}

export interface RelationshipCount {
  relationship: string;
  count: number;
}

export interface ContactSummary {
  total: number;
  relationships: RelationshipCount[];
}

export interface ErrorResponse {
  error: string;
}

export type ListContactsParams = {
search?: string;
};