# lib/api-zod/src/generated/types/contactSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RelationshipCount } from './relationshipCount';

export interface ContactSummary {
  total: number;
  relationships: RelationshipCount[];
}