# lib/api-zod/src/generated/types/emergencyContact.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyContact {
  id: number;
  name: string;
  phoneNumber: string;
  relationship: string;
  address: string;
  createdAt: Date;
  updatedAt: Date;
}