# lib/api-zod/src/generated/types/emergencyContactInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyContactInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 3
     * @maxLength 40
     */
  phoneNumber: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  relationship: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  address: string;
}