# lib/api-zod/src/generated/types/relationshipCount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelationshipCount {
  relationship: string;
  count: number;
}